rename: 내장 클래스 위치 수정
기존에는 도메인에 있었지만 도메인 안에 새로운 패키지를 생성하여 이동
